requirements.txt
Django==1.10
djangorestframework==3.4.4
mysqlclient==1.3.7
